refactor : 롤 body 가독성 개선
- LolBody -> Body 로 개선
- Body 내 코드 가독성 개선
- 게시글 필터 BoardFilter로 분리
-
