+++ src/components/main/lol/Body.d.js
@@ -1,0 +1,29 @@
+export const queueTypeList = [
+  { value: 'ALL', label: '모든 큐' },
+  { value: 'DUO_RANK', label: '솔로랭크' },
+  { value: 'FREE_RANK', label: '자유랭크' },
+  { value: 'NORMAL', label: '일반' },
+  { value: 'ARAM', label: '칼바람나락' },
+];
+
+export const tierList = [
+  { value: 'ALL', label: '모든 티어' },
+  { value: 'CHALLENGER', label: '챌린저' },
+  { value: 'GRANDMASTER', label: '그랜드마스터' },
+  { value: 'MASTER', label: '마스터' },
+  { value: 'DIAMOND', label: '다이아몬드' },
+  { value: 'PLATINUM', label: '플래티넘' },
+  { value: 'GOLD', label: '골드' },
+  { value: 'SILVER', label: '실버' },
+  { value: 'BRONZE', label: '브론즈' },
+  { value: 'IRON', label: '아이언' },
+];
+
+export const laneList = [
+  { value: 'ALL', label: '전체' },
+  { value: 'TOP', label: '탑' },
+  { value: 'JUG', label: '정글' },
+  { value: 'MID', label: '미드' },
+  { value: 'ADC', label: '원딜' },
+  { value: 'SPT', label: '서포터' },
+];
