+++ src/pages/LeagueofLegends.js
@@ -3,14 +3,14 @@ import { Fragment } from 'react';
 import MainHeader from '../components/main/MainHeader';
 import MainFooter from '../components/main/MainFooter';
 import MainBodyWrapper from '../components/main/MainBodyWrapper';
-import LolBody from '../components/main/lol/LolBody';
+import Body from '../components/main/lol/Body';
 
 const LeagueofLegends = () => {
   return (
     <Fragment>
       <MainHeader game={'lol'} />
       <MainBodyWrapper game={'lol'}>
-        <LolBody/>
+        <Body/>
       </MainBodyWrapper>
       <MainFooter />
     </Fragment>
